[NG23-136] Refactor Home and Learn views logic
