+++ lib/oli_web/live/delivery/student/index_live.ex
@@ -3,15 +3,12 @@ defmodule OliWeb.Delivery.Student.IndexLive do
 
   import OliWeb.Components.Delivery.Layouts
 
-  alias Oli.Authoring.Course.Project
-  alias Oli.Delivery.{Attempts, Sections, Metrics}
+  alias Oli.Delivery.{Attempts, Hierarchy, Metrics, Sections}
   alias Oli.Delivery.Sections.SectionCache
   alias Oli.Publishing.DeliveryResolver
   alias OliWeb.Components.Delivery.Schedule
   alias OliWeb.Common.FormatDateTime
   alias OliWeb.Delivery.Student.Utils
-
-  import Ecto.Query, warn: false, only: [from: 2]
 
   def mount(_params, _session, socket) do
     section = socket.assigns[:section]
@@ -26,17 +23,21 @@ defmodule OliWeb.Delivery.Student.IndexLive do
           Sections.get_last_open_and_unfinished_page(section, current_user_id),
           Sections.get_nearest_upcoming_lesson(section)
         ],
-        fn page ->
-          page_module_index =
-            section
-            |> get_or_compute_full_hierarchy()
-            |> find_module_ancestor(
-              page[:resource_id],
-              Oli.Resources.ResourceType.get_id_by_type("container")
-            )
-            |> get_in(["numbering", "index"])
+        fn
+          nil ->
+            nil
 
-          Map.put(page, :module_index, page_module_index)
+          page ->
+            page_module_index =
+              section
+              |> get_or_compute_full_hierarchy()
+              |> Hierarchy.find_module_ancestor(
+                page[:resource_id],
+                Oli.Resources.ResourceType.get_id_by_type("container")
+              )
+              |> get_in(["numbering", "index"])
+
+            Map.put(page, :module_index, page_module_index)
         end
       )
 
@@ -121,7 +122,10 @@ defmodule OliWeb.Delivery.Student.IndexLive do
         </div>
       </div>
 
-      <div class="w-full px-9 absolute flex-col justify-center items-start gap-6 inline-flex">
+      <div
+        :if={!is_nil(@suggested_page)}
+        class="w-full px-9 absolute flex-col justify-center items-start gap-6 inline-flex"
+      >
         <div class="text-white text-2xl font-bold leading-loose tracking-tight">
           Continue Learning
         </div>
@@ -328,111 +332,9 @@ defmodule OliWeb.Delivery.Student.IndexLive do
     FormatDateTime.to_formatted_datetime(due_date, context, format)
   end
 
-  defp find_module_ancestor(_, nil, _), do: nil
-
-  defp find_module_ancestor(hierarchy, resource_id, container_resource_type_id) do
-    case Oli.Delivery.Hierarchy.find_parent_in_hierarchy(
-           hierarchy,
-           &(&1["resource_id"] == resource_id)
-         ) do
-      %{"resource_type_id" => ^container_resource_type_id, "numbering" => %{"level" => 2}} =
-          module ->
-        module
-
-      parent ->
-        find_module_ancestor(hierarchy, parent["resource_id"], container_resource_type_id)
-    end
-  end
-
   def get_or_compute_full_hierarchy(section) do
     SectionCache.get_or_compute(section.slug, :full_hierarchy, fn ->
-      full_hierarchy(section)
-    end)
-  end
-
-  defp full_hierarchy(section) do
-    {hierarchy_nodes, root_hierarchy_node} = hierarchy_nodes_by_sr_id(section)
-
-    hierarchy_node_with_children(root_hierarchy_node, hierarchy_nodes)
-  end
-
-  defp hierarchy_node_with_children(
-         %{"children" => children_ids} = node,
-         nodes_by_sr_id
-       ) do
-    Map.put(
-      node,
-      "children",
-      Enum.map(children_ids, fn sr_id ->
-        Map.get(nodes_by_sr_id, sr_id)
-        |> hierarchy_node_with_children(nodes_by_sr_id)
-      end)
-    )
-  end
-
-  # Returns a map of resource ids to hierarchy nodes and the root hierarchy node
-  defp hierarchy_nodes_by_sr_id(section) do
-    page_id = Oli.Resources.ResourceType.get_id_by_type("page")
-    container_id = Oli.Resources.ResourceType.get_id_by_type("container")
-
-    labels =
-      case section.customizations do
-        nil -> Oli.Branding.CustomLabels.default_map()
-        l -> Map.from_struct(l)
-      end
-
-    from(
-      [s: s, sr: sr, rev: rev, spp: spp] in DeliveryResolver.section_resource_revisions(
-        section.slug
-      ),
-      join: p in Project,
-      on: p.id == spp.project_id,
-      where:
-        rev.resource_type_id == ^page_id or
-          rev.resource_type_id == ^container_id,
-      select: %{
-        "id" => rev.id,
-        "numbering" => %{
-          "index" => sr.numbering_index,
-          "level" => sr.numbering_level
-        },
-        "children" => sr.children,
-        "resource_id" => rev.resource_id,
-        "project_id" => sr.project_id,
-        "project_slug" => p.slug,
-        "title" => rev.title,
-        "slug" => rev.slug,
-        "graded" => rev.graded,
-        "intro_video" => rev.intro_video,
-        "poster_image" => rev.poster_image,
-        "intro_content" => rev.intro_content,
-        "duration_minutes" => rev.duration_minutes,
-        "resource_type_id" => rev.resource_type_id,
-        "section_resource" => sr,
-        "is_root?" =>
-          fragment(
-            "CASE WHEN ? = ? THEN true ELSE false END",
-            sr.id,
-            s.root_section_resource_id
-          )
-      }
-    )
-    |> Oli.Repo.all()
-    |> Enum.map(fn node ->
-      numbering = Map.put(node["numbering"], "labels", labels)
-
-      Map.put(node, "uuid", Oli.Utils.uuid())
-      |> Map.put("numbering", numbering)
-    end)
-    |> Enum.reduce({%{}, nil}, fn item, {nodes, root} ->
-      {
-        Map.put(
-          nodes,
-          item["section_resource"].id,
-          item
-        ),
-        if(item["is_root?"], do: item, else: root)
-      }
+      Hierarchy.full_hierarchy(section)
     end)
   end
